refactor(MemberException, MemberErrorCode) : 패키지 경로 수정으로 인한 리팩토링
