+++ src/main/java/org/noostak/auth/domain/vo/Code.java
@@ -2,8 +2,8 @@ package org.noostak.auth.domain.vo;
 
 import jakarta.persistence.Embeddable;
 import lombok.EqualsAndHashCode;
-import org.noostak.member.common.MemberErrorCode;
-import org.noostak.member.common.MemberException;
+import org.noostak.member.common.exception.MemberErrorCode;
+import org.noostak.member.common.exception.MemberException;
 
 @Embeddable
 @EqualsAndHashCode
