+++ src/main/java/org/noostak/member/common/exception/MemberErrorCode.java
@@ -1,4 +1,4 @@
-package org.noostak.member.common;
+package org.noostak.member.common.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
