+++ src/main/java/org/noostak/member/common/exception/MemberException.java
@@ -1,4 +1,4 @@
-package org.noostak.member.common;
+package org.noostak.member.common.exception;
 
 public class MemberException extends RuntimeException {
     public MemberException(MemberErrorCode errorCode) {
